added small formatting and syntax style changes. Updated ruff and mypy rules.
