+++ src/middleware/logging.py
@@ -1,4 +1,3 @@
-import json
 import sys
 
 from fastapi import Request
